Display human-readable names in model mappings.
